refactor: 지도 하위 페이지 모델 정리 및 공통 타입 엔티티로 분리
