+++ apps/web/src/entities/place/index.ts
@@ -1,4 +1,4 @@
 export { getCategoryLabel } from './lib/category-label';
 export type { PlaceCategory, PlaceCategoryValue } from './model/place-category';
 export { PLACE_CATEGORIES } from './model/place-category';
-export type { Place, PlaceLayer } from './model/types';
+export type { MapSortType, Place, PlaceLayer } from './model/types';
