+++ apps/web/src/views/map-search/model/use-map-search-query.ts
@@ -4,7 +4,14 @@ import { useQuery } from '@tanstack/react-query';
 
 import { clientApi } from '@/shared/api/client-api';
 
-import { type MapSearchPlace } from './types';
+export type MapSearchPlace = {
+  placeId: number;
+  placeName: string;
+  address: string;
+  latitude: number;
+  longitude: number;
+  lastStudyDate: string;
+};
 
 function fetchMapSearchPlaces(keyword: string) {
   return clientApi.get<MapSearchPlace[]>(
